Move reflect transaction class to apply_transaction package
